Tenancy support (#16)
* First cut - custom tenancy fields.

- New settings record without tenancy details defaults to Constants.
- New UI elements to support custom tenancy details
- Sandbox connection factory refactored and documented, with api endpoint pulled in from MindgardSettings as opposed to a constant.
- Other javadocs

* - UI has login button and test configuration tabs
- Login flow removed from the initialization of the extension
- Removed MindgardSettingsManager as not needed
- Settings passed in to MindgardAuthentication, because the api endpoint for tests is pulled from the url field on the record

* UI refactoring

* MindgardSettingsManager re-introduced and used in appropriate places.

Co-authored-by: Rob Heath <robert.heath@mindgard.ai>

* Fix validation of project ID

Modify the validation logic to throw an error, and catch it in the
settings.

@Co-authored-by: Andrew Cook <andrew.cook@mindgard.ai>

* Use MindgardToken to store the auth token

The existing method of storing tokens is to write to token.txt in the
MindgardAuthentication model using the MindgardSettings file getter.
This change moves the token into its own class, controlled by the
MindgardSettingsManager, which allows simpler reading/writing and
persistence of state wherever the manager is loaded.

* Modify MindgardAuthenticationTest not to write to disk

This was overwriting the settings for Burp every time the tests were
run.

* Add validLogin boolean method to settings manager

This will allow us to control flow through Mindgard's UI by determining
whether or not a token is present, and if so, whether it is for the
currently defined Mindgard base URL.

I have also removed the redundant functionality of passing a filename
into the loadOrCreate method.

* Add visual indicator of login status

This adds a text indicator of the login status for a given tenant. It
also flashes a warning if you try to use the generator without being
logged in.

* Quality of life improvements for logging in: timeouts, warnings

This modifies the login behaviour in a few key ways:
1. It adds a one minute timeout for a user authing. If a user does not
   auth before that time, a warning is shown and the process resets.
   This prevents Burp from hanging indefinitely if the customer
   rejects the key request.
2. It adds a warning if there are changes to settings that have not been
   saved when a login is attempted. This prevents a user from changing
   the server URL and then attempting to sign in with the server address
   incorrectly set.

* Add negative space above settings panel objects

This makes the settings panel look more sensible.

---------

Co-authored-by: Rob Heath <robert.heath@mindgard.ai>